feat: backend-service-refactor-service-package-003 - Extract hosted workers and hosted poller package
